feat(crypto): rename user file to api
