+++ packages/crypto/src/index.ts
@@ -2,7 +2,7 @@ import {definePackage} from '@alwatr/logger';
 
 export * from './hash.js';
 export * from './token.js';
-export * from './user.js';
+export * from './api.js';
 export * from './type.js';
 export * from './pre-config.js';
